refactoring to avoid key containing change specific info
